refactor: 유틸함수 테스트끼리 묶어서 관리
