+++ frontend/src/utils/getInfoByTeamPlaceId.ts
