+++ frontend/src/utils/test/parseDate.test.ts
@@ -1,4 +1,4 @@
-import { parseDate } from './parseDate';
+import { parseDate } from '../parseDate';
 
 describe('parseDate', () => {
   it('인자로 전달받은 Date객체를 년, 월, 일, 요일을 갖는 객체로 반환한다.', () => {
